Move Listing to models and ListingIpfsStore to services
