+++ src/services/listing-service.js
@@ -13,45 +13,6 @@ ajv.addMetaSchema(require('ajv/lib/refs/json-schema-draft-06.json'))
 ajv.addSchema(schemaV1)
 
 
-//
-// Listing is the main interface exposed by Origin Protocol to access data from a listing.
-//
-class Listing {
-  /**
-   * A Listing is constructed based on its on-chain and off-chain data.
-   * #param {string} listingId - Unique listing ID.
-   * @param {Object} chainListing - Listing data from the blockchain.
-   * @param {Object} ipfsListing - Listing data from IPFS.
-   */
-  constructor(listingId, chainListing, ipfsListing) {
-    this.id = listingId
-    // FIXME(franck): Exposing directly the chain data will make it difficult
-    // to support backward compatibility of the Listing interface in the future. We should
-    // select and possibly abstract what data from the chain gets exposed.
-    Object.assign(this, ipfsListing, chainListing)
-  }
-
-  get unitsSold() {
-    // Lazy caching.
-    if (this._unitsSold !== undefined) {
-      return this._unitsSold
-    }
-    this._unitsSold = Object.keys(this.offers).reduce((acc, offerId) => {
-      if (this.offers[offerId].status === 'created') {
-        return acc + 1
-      }
-      // TODO: we need to subtract 1 for every offer that is canceled
-      return acc
-    }, 0)
-    return this._unitsSold
-  }
-
-  get unitsRemaining() {
-    // Should never be negative.
-    return Math.max(this.unitsTotal - this.unitsSold, 0)
-  }
-}
-
 /**
  * Returns an adapter based on a schema version.
  * @param {string} schemaVersion
@@ -125,7 +86,7 @@ class SchemaAdapterV1 {
 //
 // ListingIpfsStore exposes methods to read and write listing data from/to IPFS.
 //
-class ListingIpfsStore {
+export class ListingIpfsStore {
   constructor(ipfsService) {
     this.ipfsService = ipfsService
   }
@@ -218,9 +179,4 @@ class ListingIpfsStore {
     const ipfsHash = await this.ipfsService.saveObjAsFile(ipfsData)
     return ipfsHash
   }
-}
-
-module.exports = {
-  Listing,
-  ListingIpfsStore
 }
