+++ test/listing-service.test.js
@@ -1,35 +1,12 @@
 import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
-chai.use(chaiAsPromised)
-
-const expect = chai.expect
-
-import { Listing, ListingIpfsStore} from '../src/resources/listing.js'
 import sinon from 'sinon'
+
+import { ListingIpfsStore} from '../src/services/listing-service'
 import goodListing from './data/listing-valid'
 
-
-describe('Listing', () => {
-
-  it(`unitsRemaining should equal to unitsTotal when no offer`, () => {
-    const chainListing = { offers: [] }
-    const ipfsListing = { unitsTotal: 10 }
-    const listing = new Listing('FakeID', chainListing, ipfsListing)
-    expect(listing.unitsRemaining).to.equal(10)
-    expect(listing.unitsSold).to.equal(0)
-  })
-
-  it(`unitsRemaining should be unitsTotal - unitsSold`, () => {
-    const chainListing = { offers: {
-      offerId1: { status: 'created' },
-      offerId2: { status: 'accepted'} } }
-    const ipfsListing = { unitsTotal: 10 }
-    const listing = new Listing('FakeID', chainListing, ipfsListing)
-    expect(listing.unitsRemaining).to.equal(9)
-    expect(listing.unitsSold).to.equal(1)
-  })
-
-})
+chai.use(chaiAsPromised)
+const expect = chai.expect
 
 
 describe('ListingIpfsStore load', () => {
